TestingCamera: Add camera testing application for PDK users
TestingCamera is an app for testing all the features of the Android
camera API, targeted at platform engineers. Unlike the standard camera
app, it is not meant to be an end-user application, and focuses more
on exposing all features and options in the API, instead of on ease of
use.

This initial checking supports camera selection, preview and still
image capture only.

Change-Id: I303e74ba3f33f3b41ba5219e1747e8fae3a2cd56